refactor(rusty-stack): unify into a single `rusty-stack` binary
The `rusty` and `rusty-stack` binaries were redundant: `rusty-stack` (main.rs)
was a 3-line TUI wrapper while `rusty` (bin/rusty.rs) was the full CLI + TUI.
Collapsed to ONE binary, `rusty-stack`, containing the full CLI (all subcommands
+ TUI on no args):

- Moved src/bin/rusty.rs -> src/main.rs (full CLI is now the single entry).
- Removed the separate `[[bin]] rusty` target; deleted src/bin/.
- clap command name `rusty` -> `rusty-stack` (--version/help now match).
- Self-upgrade `expected_binary_name()` `rusty` -> `rusty-stack` so `rusty-stack
  upgrade` finds the right asset in the release tarball.

Backward compat: `rusty` is provided as a symlink to `rusty-stack` at install
time, so existing `rusty <cmd>` usage and docs keep working (one binary, alias).

Validated: cargo fmt clean, clippy --workspace --all-targets = 0 warnings,
release builds a single `rusty-stack` binary; all 8 subcommands + TUI present;
`--version` reports `rusty-stack 0.3.1`.

NOTE: release CI must publish a `rusty-stack` asset (not `rusty`) to match
expected_binary_name; follow-up when cutting the next release.
